Updating

+++ server.py
@@ -1,8 +1,11 @@
-from flask import Flask, render_template, request, redirect, url_for
+from flask import Flask, render_template, request, redirect, url_for, flash
 import data_handling as data_handling
 import util as util
+from werkzeug.utils import secure_filename
+import os
 
 app = Flask(__name__)
+app.config['UPLOAD_FOLDER'] = data_handling.UPLOAD_FOLDER
 
 @app.route("/")
 def hello():
@@ -18,17 +21,30 @@ def list():
     question_list_to_show = util.get_question_list_with_real_time(question_list)
     return render_template("list.html", question_list=question_list_to_show, headers=headers)
 
-@app.route("/question/<question_id>")
+@app.route("/question/<question_id>", methods=["POST", "GET"])
 def display(question_id):
-    question_to_display, headers = util.get_questions_to_display(question_id)
-    final_answer_list = util.get_answer_to_display(question_id)
+    question_to_display = util.get_questions_to_display(question_id)
+    headers = data_handling.get_headers_answers()
+    answer_list = util.get_answer_to_display(question_id)
+    order_by = request.args.get('order_by', 'vote_number')
+    order_direction = request.args.get('order_direction', 'asc')
+    answer_list_to_show = util.sort_table(order_by, order_direction, answer_list)
+    final_answer_list = util.get_question_list_with_real_time(answer_list_to_show)
     return render_template("display.html", question=question_to_display, answers=final_answer_list, headers=headers)
+
 
 @app.route("/add_question", methods=["POST", "GET"])
 def add_question():
     if request.method == "POST":
+        if "file" in request.files:
+            file = request.files['file']
+            filename = secure_filename(file.filename)
+            if filename:
+                file.save(os.path.join(app.config["UPLOAD_FOLDER"], filename))
+            else:
+                filename = False
         new_question_input = dict(request.form)
-        data_handling.save_question(new_question_input)
+        data_handling.save_question(new_question_input, filename)
         return redirect(f"/question/{new_question_input['id']}")
     return render_template("add_question.html")
 
@@ -36,8 +52,16 @@ def add_question():
 @app.route("/question/<question_id>/new-answer", methods=["GET", "POST"])
 def post_an_answer(question_id):
     if request.method == "POST":
+        if "file" in request.files:
+            file = request.files['file']
+            filename = secure_filename(file.filename)
+            if filename:
+                filename = str(question_id) + "_" + filename
+                file.save(os.path.join(app.config["UPLOAD_FOLDER"], filename))
+            else:
+                filename = False
         answer = dict(request.form)
-        data_handling.save_answer(question_id, answer)
+        data_handling.save_answer(question_id, answer, filename)
         return redirect( f"/question/{answer['question_id']}")
     return render_template("post_an_answer.html", question_id=question_id)
 
@@ -49,34 +73,38 @@ def edit_question(question_id):
         new_data = dict(request.form)
         util.edit_single_question(question, new_data)
         data_handling.overwrite_question(question)
-        return redirect( f"/question/{question_id}")
+        return redirect(f"/question/{question_id}")
     return render_template("edit_question.html", question=question, question_id=question_id)
 
 
 @app.route("/answer/<answer_id>/delete", methods=["GET", "POST"])
 def delete_answer(answer_id):
     answer = data_handling.get_answer_by_id(answer_id)
-    if request.method == "POST":
-        if 'yes_button' in request.form:
-            data_handling.delete_answer(answer)
-            return redirect(f"/question/{answer['question_id']}")
-        else:
-            return redirect(f"/question/{answer['question_id']}")
-
-    return render_template("delete_answer.html", answer_id=answer_id)
+    if answer:
+        if request.method == "POST":
+            if 'yes_button' in request.form:
+                data_handling.delete_answer(answer)
+                return redirect(f"/question/{answer['question_id']}")
+            else:
+                return redirect(f"/question/{answer['question_id']}")
+        return render_template("delete_answer.html", answer_id=answer_id)
+    else:
+        return render_template("wrong_answer_id.html")
 
 
 @app.route("/question/<question_id>/delete", methods=["GET", "POST"])
 def delete_question(question_id):
     question = data_handling.get_question_by_id(question_id)
-    if request.method == "POST":
-        if 'yes_button' in request.form:
-            data_handling.delete_question(question)
-            return redirect("/list")
-        else:
-            return redirect(f"/question/{question['id']}")
-
-    return render_template("delete_question.html", question_id=question_id)
+    if question:
+        if request.method == "POST":
+            if 'yes_button' in request.form:
+                data_handling.delete_question(question)
+                return redirect("/list")
+            else:
+                return redirect(f"/question/{question['id']}")
+        return render_template("delete_question.html", question_id=question_id)
+    else:
+        return render_template("wrong_question_id.html")
 
 
 @app.route("/question/<question_id>/vote_up", methods=["GET", "POST"])
